LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-same-in-same-self-allow-expected.txt

A 'frame-ancestors' CSP directive with a value 'same' should block render in same-origin nested frames. This test PASSED if you see content in the inner-most nested iframe below. Otherwise, it FAILED.



--------
Frame: '<!--frame1-->'
--------
Testing a same-origin child with a policy of "'self'" nested in a same-origin parent.



--------
Frame: '<!--frame2-->'
--------
This is an IFrame sending a Content Security Policy header containing "frame-ancestors 'self'".
